requirements.txt
psycopg2
flask
flask-sqlalchemy
python-dotenv
python-dateutil
uwsgi
gunicorn